requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
dnspython==1.16.0
email_validator==2.1.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.26.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
motor==3.3.2
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.9.15
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.5.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.14.6
pymongo==4.6.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==20.8
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
